add webelements to Products, rename Products to ProductsPage, add click methods to ProductsPage
